♻️ 6DMWEY task: split runner task-state helpers
